[native] Add active state to `ColorSelector`
Summary:
Part of ongoing redesign of `ColorPicker` component on `native` to match the `ColorSelector` component on `web`. Depends on D3798.

Adds active state to `ColorSelector` by adding styling and a checkmark to the `ColorSelectorButton` that is currently selected as the thread color.

Will add `onPress` handling and updating the thread color in a later diff, as outlined in the plan on this Linear issue.

Test Plan: Tested on iOS 15.4 and works as expected:

Reviewers: atul

Subscribers: ashoat, palys-swm, Adrian, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D3868