[web] Introduce `AddMembersModal` component
Summary:
Introduce "AddMembersModal", that is used for searching and selecting users to add to a thread.
There is no logic for searching users, and for dispatching server action, added in this diff yet - they will be added in the following diffs.

Test Plan: The modal will be displayed after pressing "Add members" button in MembersModal - after introducing the action in the following diffs.

Subscribers: ashoat, palys-swm, Adrian, atul, karol-bisztyga, benschac, yayabosh

Differential Revision: https://phabricator.ashoat.com/D3742