[web] [feat] [ENG-991] add show promote thread logic to web
Summary:
add logic to promote chats that are promotable

no promote
{F43080}

promote
{F43081}

Test Plan: click on a sidebar that meets promotable criteria, promote option should be shown to you as an end user. Go to a non-promotable sidebar, it shouldn't have the promote option. logic was moved into the `lib/hooks/promote-thread` hook in a previous commit D3807. No new logic was introduced.

Reviewers: atul, palys-swm, def-au1t, yayabosh

Reviewed By: palys-swm

Subscribers: ashoat, Adrian, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D3810