[web] Display unfriend action menu
Summary:
When an edit button is clicked we display a menu where a user can unfriend another user
Depends on D4006

Test Plan:
Click an edit button and check if menu is displayed.
Click outside of it and check if it closes.
Open it and click another edit button - verify that only one menu is opened at a time.
Open menu and check if `unfriend` action works.

Reviewers: benschac, atul, def-au1t

Subscribers: ashoat, Adrian, yayabosh

Differential Revision: https://phabricator.ashoat.com/D4007