[landing] Add thumbnail and description for landing page when shared
Summary: Add image thumbnail and website description (from Twitter bio) for the landing page so it appears in Twitter Cards/Facebook messages/posts when shared. Facebook called this "Sandcastle."

Test Plan: N/A, as of now. There is no staging environment with a public URL so not sure how to test these changes yet.

Reviewers: atul

Subscribers: