[web] Fix `See more...` button alignment regression
Summary: Noticed a regression in the `See more...` sidebar button's
alignment with the sidebars. Fixed this by adjusting the `padding-left`.

As an aside, because the `ChatThreadList` component is being actively
worked on, I suspect the alignment may be messed up in the future if the
arrow/sidebar get bigger or smaller. Therefore, we should probably pull
in the `padding` from some central place so we don't introduce new
regressions when modifying the styling of one part of the component.

Test Plan: Tested on Chrome/Safari and works as expected.

Reviewers: atul

Subscribers: