[CI] Include Buildkite pipelines in repo
Summary:
We were previously using a mix of the "Legacy" and "YAML" steps for our Buildkite pipelines.

Now that we've moved all of the pipelines over to the "YAML" approach, we can check the files into the repo.

Once they're checked into the repo, we can "pull them" into Buildkite by replacing the contents of the pipelines in the web interface with something like:

```
steps:
   - command: buildkite-agent pipeline upload
     label: ":ios:"
```

Test Plan:
These are the steps we are using already with the Buildkite CI. Ran them a few times after converting the steps to YAML and things continue to work as expected.

Reviewers:
ashoat, palys-swm