[web] Move padding from `modal_body` to `tab_body` in `ThreadSettingsModal`
Summary:
We want each of the `ThreadSettingsBlahTab` components to have some padding so their contents aren't right against the edge of the surrounding Modal. However, we **do** want the Tabs to extend all the way to the edges and be closer to the Modal header.

We can't put these divs inside `<Tabs.Container>` because it expects its children to be typed a certain way. As a result we have to copy/paste the `<div className={css.tab_body}>` tag for every tab. We **could** move this padding to the children Tab components themselves, but setting the padding in `ThreadSettingsModal` seems like the right layer of abstraction to me.

Test Plan:
Before:
After:

Reviewers:
ashoat, palys-swm