[web] Move `modal_form_error` within `tab_body`
Summary:
We want the error message to appear "within" the `tab_body` so it isn't pushed to the outer edge of the Modal. This required "copy/pasting" the error message div three times... couldn't think of a way around this because only `Tabs.Item` components can be nested within `Tabs.Container`.

Note: As an aside, I think displaying the error message within the "Save" button could be cleaner and keep the layout simpler. We do something like this on `landing` with the `SubscriptionForm` component. Not saying we should prioritize this right now, but just thought it might look cleaner. Might looks something like:

---

Depends on D3933

Test Plan:
Here's what it looks like:

Before:

After: