[Identity] Token constructor
Summary:
new function to generate a Token. we use the rand crate to generate a random string for the token, and chrono to get the creation time. I used len 512 for the string because that's what OAuth uses for access tokens.

@jimpo, is this the best way to generate a random String in Rust?

Depends on D3957

Test Plan: Called constructor and verified the contents of the Token

Reviewers: jimpo, karol-bisztyga, palys-swm

Subscribers: atul