[services] Backup/Blob - Split aws tools into two tool sets
Summary:
Depends on D3913

We can split `AwsTools` into `DynamoDBTools` and `S3Tools`. I did it because the tools for the S3 are only used in the blob service and the tools for dynamo DB are used in both services. Therefore we could move dynamo DB tools to the mutual space.

Test Plan:
```
cd services
yarn run-local-cloud
yarn test-backup-service-dev-mode
yarn test-blob-service-dev-mode
```

Reviewers: palys-swm

Subscribers: ashoat, palys-swm, Adrian, atul, yayabosh

Differential Revision: https://phabricator.ashoat.com/D3942