[services] Merge .dockerignore files to the one in the root folder
Summary: Merging docker ignore file from the backup service into the one in the root directory of the project.

Test Plan:
Hard to come up with a certain plan here.
Services have to build, as usual, this is a build change.
To better check this, you'd have to manually log into the shell of a certain service and list files to see if the files listed in docker ignore are not copied to the container. You can add `ls -l` command to the `CMD` command in `Dockerfile`s

Reviewers: palys-swm, ashoat

Subscribers: ashoat, palys-swm, Adrian, atul, yayabosh

Differential Revision: https://phabricator.ashoat.com/D3967