[services] Fix dockerignore file
Summary:
Depends on D3967

I wanted to make this file more readable and also I added an entry for the blob service

Test Plan:
Hard to come up with a certain plan here.
Services have to build, as usual, this is a build change.
To better check this, you'd have to manually log into the shell of a certain service and list files to see if the files listed in docker ignore are not copied to the container. You can add `ls -l` command to the `CMD` command in `Dockerfile`s

Reviewers: palys-swm, ashoat

Reviewed By: ashoat

Subscribers: ashoat, palys-swm, Adrian, atul, yayabosh

Differential Revision: https://phabricator.ashoat.com/D3968