[native] [fix] [ENG-1123] update message chat bubble design
Summary: https://linear.app/comm/issue/ENG-1123/update-chat-bubble-design-to-match-figma, the message bubble needs to be updated so that reply-engagement component isn't too cramped next to the text in the message bubble. Additoinally, these changes are in line with main stream messaging design.

Test Plan: review figma, and make sure this looks like the design.

Reviewers: @atul, @def-au1t, @palys-swm

Subscribers: