[iOS] Fix `CryptoTest.mm`
Summary:
These tests weren't building, and as a result weren't able to be run, etc.

Goal is to eventually add these tests to the iOS CI, but ran into some issues with simulator architectures and whatnot so will handle that in a subsequent diff.

Test Plan: Tests run and pass (from Xcode GUI):

Reviewers: palys-swm, def-au1t, benschac

Subscribers: ashoat, Adrian, yayabosh

Differential Revision: https://phabricator.ashoat.com/D4020