[keyserver] Use importJSON to import olm_config.json
Summary:
It seems that this function is not used anywhere, so I am hoping it's okay to switch to an async function.

(The reason for this is that I want to use `importJSON` so I can later refactor all configs so they can be specified by environmental variables in the Docker context.)

Depends on D4024

Test Plan: @marcinwasowicz, any advice on how to test?

Reviewers: marcinwasowicz, palys-swm

Subscribers: