[web] Make `valid` prop of `EnumSettingsOptionInfo` optional
Summary:
Make `valid` prop of `EnumSettingsOptionInfo` optional.

Before, when this component was only being used by `NotificationsModal`, it was being passed `description` which was an array of [text, isValid] tuples.

Based on the `isValid` value, the `text` statement would be styled as being factual or false.

However, when we use this component elsewhere--for example in `ThreadSettingsPrivacyTab`, we're not going to have a list of true/false statements. Rather, we're going to have a brief text description.

This diff lets us include that brief text description without the accompanying styling (icon + strikethrough (or absense of strikethrough)).

In subsequent diffs `EnumSettingsOption` will be refactored further to make things cleaner. Specifically, changing the `description` type and variable names.

---

Depends on D4046

Test Plan:
1. Manually modify `description` (array of tuples) such that the second value is undefined/null.
2. Observe that the `EnumSettingsOption` appears as expected without check/cross icon or any strikethroughs.

Reviewers: def-au1t, palys-swm, benschac

Subscribers: ashoat, Adrian, yayabosh

Differential Revision: https://phabricator.ashoat.com/D4047