[native] Update `sqlite3` entry to point to latest release
Summary:
Context: https://linear.app/comm/issue/ENG-373/update-sqlite3-entry-in-nativepackagejson-on-new-release

TLDR: the package wasn't being maintained and there hadn't been a release with the changes we needed... so we specified a commit here. now we can just point to the latest release which includes the changes we need.

Test Plan:
CI

Reviewers:
def-au1t, palys-swm, ashoat!, benschac