[web] Fix thread description and color modification in private thread
Summary:
Fix issue where user is unable to modify the thread description
and color in their private thread. The thread name is still unmodifiable
(following the introduction of the `EDIT_DESCRIPTION` and `EDIT_COLOR`
permissions, but not the `EDIT_THREAD_NAME` permission in D1980), but the
thread description and color can be modified.

Test Plan: Tested on Chrome/Safari.

Reviewers: atul

Subscribers: ashoat, palys-swm, Adrian

Differential Revision: https://phabricator.ashoat.com/D4073