[web] Introduce danger zone route
Summary:
Add a new route for danger zone. It looks like this is the first time we introduce a route consisting of two words, but using hyphens sounds like a better approach compared to underscores.

Test Plan:
Open settings and check if two nav items are displayed. Try switching between them - always the active should be highlighted. When danger zone is opened, an empty page should be displayed.

Reviewers: def-au1t, atul, benschac