[keyserver] Make importJSON check env variables
Summary:
As we Dockerize the keyserver, we need to make it possible to define config via environmental variables. This is a standard practice in Docker for specifying config/secrets.

Currently we specify config/secrets to Node via JSON files that get imported directly. In previous diffs in the stack, I've refactored all of these codepaths to go through `importJSON`. In this diff, I make it possible for `importJSON` to get the config from environmental variables rather than JSON files.

Depends on D4101

Test Plan: I tested this diff in combination with the next one. See the test plan there

Reviewers: palys-swm, atul

Subscribers: Adrian, yayabosh

Differential Revision: https://phabricator.ashoat.com/D4102