[services] Tunnelbroker - Introducing `deleteQueueIfEmpty` method for DeliveryBroker
Summary:

Following the comment on D3833 introducing the `deleteQueueIfEmpty` method. The method purpose is if `clientDeviceID` messages queue is empty we don't
need to store `folly::MPMCQueue` for it and need to free memory to fix possible 'ghost' queues.

Linear task: ENG-1146

Test Plan:

Run `yarn run-tunnelbroker-service`, service is built and run after.

Reviewers:

Subscribers: