[web] Rename `inputDisabled` to `threadSettingsOperationInProgress` in `ThreadSettingsTab`s
Summary:
Renamed `inputDisabled` prop/variable in
`ThreadSettingsGeneralTab`, `ThreadSettingsPrivacyTab`,
`ThreadSettingsDeleteTab`, and `ThreadSettingsModal` to
`threadSettingsOperationInProgress`. This is because the `inputDisabled`
prop is determined by

1. Determining if there is a change in progress
2. Checking the current tab and determining if the current thread has
that tab's permissions

Thus, if `inputDisabled` is true, a thread settings operation is in
progress or the current thread does not have the current tab's
permissions. So `inputDisabled` is renamed to more accurately reflect the
cases in which it is true.

Test Plan:
Used find-and-replace to rename all instances of
`inputDisabled` in `ThreadSettings` tabs and modal. Also opened each tab
to make sure everything looked as expected.

Reviewers: atul

Subscribers: ashoat, palys-swm, Adrian

Differential Revision: https://phabricator.ashoat.com/D4105