[web] Restyle delete account modal
Summary:
Try to improve the style of the modal and make it consistent with e.g. password modal. There are no designs for this modal, so any advices will be really appreciated.

Depends on D4118

Test Plan:
Open the modal and check if it is displayed correctly.
Click `Delete Account` and check if an `unknown error` is displayed.
Type an incorrect password, click `Delete Account` and check if `wrong password` error is displayed.
Type a correct password, click the button and verify that the account was deleted.

Reviewers: def-au1t, atul, benschac

Subscribers: ashoat, Adrian, yayabosh

Differential Revision: https://phabricator.ashoat.com/D4119