[native][lib] "Permanently" make changes for release
Summary:
We made changes to `file-utils:mediaConfig` and `media-gallery-keyboard:fetchPhotos` to accomodate the video messages work when it was in progress.

The change in `file-utils` was to treat animated GIFs as videos.

The change in `media-gallery-keyboard` was to include videos in the media selection gallery.

Since the functionality to actually send video messages was not yet ready, these changes were reverted before making a release... and then brought back after making a release.

However, this adds some overhead to the release process (two extra commits) and video message work is still a ways out.

We decided in https://linear.app/comm/issue/ENG-170/daily-iosandroid-releases that it was reasonable to make the changes "permanently."

I left comments in the code with the steps to "re-make" the changes when video message work starts back up.

Test Plan:
NA

Reviewers:
def-au1t, palys-swm, ashoat