[web] Add a rectangle next to settings icon when settings page is opened
Summary:
When settings page is opened we should mark it by adding a rectangle next to the icon in bottom-left corner. https://www.figma.com/file/a1nkbWgbgjRlrOY9LVurTz/Comm-%2F-Desktop-app?node-id=1272%3A88750

Depends on D4121

Test Plan: Open chat and check if the rectangle is not displayed. Open settings and check if the rectangle is there (on both account and danger-zone pages).

Reviewers: def-au1t, atul, benschac

Subscribers: ashoat, Adrian, yayabosh

Differential Revision: https://phabricator.ashoat.com/D4124