[services] Tunnelbroker - Check for an empty string in ConfigManager
Summary:
This diff fixes a bug when the option in config file provided, but has an empty value the app throwing an following error:
`what():  basic_string::_M_construct null not valid`.
To fix this we will check the option value not to be empty or throw an appropriate error.

Linear task: ENG-1059

Test Plan:
In the config file located at `services/tunnelbroker/tunnelbroker.ini` change the option value to an empty.
Run service using the command `yarn run-tunnelbroker-service` and expect the 'Error: config parameter X can not be empty' error.

Subscribers: ashoat, palys-swm, Adrian, atul, karol-bisztyga, yayabosh

Differential Revision: https://phabricator.ashoat.com/D4137