[service][identity] helper function to handle pake registration
Summary:
This helper function takes the PAKE registration bytes from the gRPC request and deserializes them. Then it generates a message to return to the client and sends the message to the bidi stream via the mpsc Sender.

I've also added an Error enum to this module. Thought it made sense to include both in a single diff for context.

Depends on D3962

Test Plan: confirmed that the Receiver of the mpsc channel received a RegistrationResponse

Reviewers: jimpo, karol-bisztyga, atul, palys-swm

Subscribers: ashoat, Adrian, yayabosh

Differential Revision: https://phabricator.ashoat.com/D4140