[native] [fix] [ENG-1199] fix crashing app
Summary: The application is currently in a broken state. If you open the see more sidebar modal then close the app, open the app the app will stay in a loading state indefinitely. Adding @ashoat since this is a pretty urgent issue that needs to be fixed ASAP.

Test Plan: try the above locally to confirm app is broken. Patch this diff and try the same steps again, the app should load as expected. Additionally, try search sidebars functionality to make sure it's still working as expected.

Reviewers: atul, palys-swm, ashoat, def-au1t

Subscribers: Adrian, yayabosh, ashoat

Differential Revision: https://phabricator.ashoat.com/D4143