[services] Tunnelbroker - Fix arguments table format for new amqp queue
Summary:
This diff fixes a following bug:
When the queue for the tunnelbroker-id is not created (clean rabbitMQ instance) there is an error when creating a new one:

```
AMQP: Queue creation error: PRECONDITION_FAILED - inequivalent arg 'x-expires' for queue 'tunnelbroker1' in vhost '/'.
```

We must remove converstion to the `string` when initializing the queue attribute. That fixes the error.

Test Plan:
On the master branch where rabbitMQ Docker container instance was not initialized before (was clean) run `yarn run-tunnelbroker-service-dev-mode` and
get an queue creation error (which is described above in description).

Apply the diff patch and re-run `yarn run-tunnelbroker-service-dev-mode`. The rabbitMQ queue will be created successfully.

Subscribers: ashoat, palys-swm, Adrian, atul, karol-bisztyga, yayabosh

Differential Revision: https://phabricator.ashoat.com/D4168