[native] Update `threadStore` with `threadStoreOps` in `fixUnreadActiveThread`
Summary:
Based on months with `assertThreadStoreThreadsAreEqual` enabled on staff builds, I think it's safe to conclude that processing the `threadStore` with the `threadStoreOps` leads to an equivalent result. There's no need to create an updatedThreadStore twice, so we're removing the non-ops approach.

---

Depends on D4193

Test Plan: `assertThreadStoreThreadsAreEqual` hasn't flagged anything since it's been included in staff builds

Reviewers: def-au1t, palys-swm

Subscribers: ashoat, Adrian, yayabosh

Differential Revision: https://phabricator.ashoat.com/D4194